+++ satori/rtm/client.py
@@ -89,7 +89,8 @@ Syntax
         """
 
         assert endpoint
-        assert endpoint.startswith('ws://') or endpoint.startswith('wss://')
+        assert endpoint.startswith('ws://') or endpoint.startswith('wss://'),\
+            'Endpoint must start with "ws(s)://" but "%s" does not' % endpoint
 
         self._queue = queue.Queue(maxsize=max_queue_size)